1. 修改功能: Domain統一用Long 2. 修改功能: Dd301.dd300id中宣告關連 3. 其他: PA統一用Decimal

+++ grails-app/domain/Nw300.groovy
@@ -16,7 +16,7 @@ class Nw300 {
 
 
 		//---2.VERSION  
-		Integer version = 1
+		Long version = 1
 		//---3.建立人  
 		String manCreatedId
 		//---4.建立時間
